+++ SlashBar/Modules/Weather/WeatherHour.cs
@@ -1,0 +1,8 @@
+namespace SlashBar.Modules.Weather;
+
+public sealed class WeatherHour {
+
+    public string Label { get; init; } = "";
+    public string Temperature { get; init; } = "";
+    public Uri Icon { get; init; } = null!;
+}